Small fixes and move tests
